*: BB_BANNER -> bb_banner (it is not a const or #define)! correct_password: explain in detail why it is ok to use bb_banner fsck_minix: make it print bb version, not it's own (outdated/irrelevant) one
Marginal size difference:
   text    data     bss     dec     hex filename
 679119    2700   15632  697451   aa46b busybox_old
 679091    2700   15632  697423   aa44f busybox_unstripped

+++ util-linux/mkfs_minix.c
@@ -94,7 +94,7 @@ struct globals {
 	int dev_fd;
 
 #if ENABLE_FEATURE_MINIX2
-	int version2;
+	smallint version2;
 #define version2 G.version2
 #endif
 	char *device_name;
